Fix Service and facts ACL
Prior to this commit the /facts entry for auth.conf did not allow
child masters to replace facts. We now overirde the default with a regex
that allows all masters to submit thier own facts. PuppetDB should
supercede this behavior.

Also added a Service resource for pe-httpd, this gets notified when
the puppetmaster.conf file is modifed